requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
channels>=4.1,<5.0
cloudinary>=1.41.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
